refactor: 페이지 네이밍 수정
